backend/requirements.txt
alabaster==0.7.13
anyio==3.6.2
asgiref==3.6.0
astatine==0.3.2
astor==0.8.1
astpretty==3.0.0
astroid==2.15.4
asttokens==2.2.1
attrs==23.1.0
autoflake==1.7.8
Babel==2.12.1
bandit==1.7.5
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cognitive-complexity==1.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.5
cryptography==40.0.2
darglint==1.8.1
dill==0.3.6
distlib==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
dlint==0.14.1
doc8==1.1.1
docformatter==1.6.5
docopt==0.6.2
docutils==0.19
domdf-python-tools==3.6.1
drf-yasg==1.21.4
eradicate==2.2.0
exceptiongroup==1.1.1
fastapi==0.95.1
filelock==3.12.0
flake8==5.0.4
flake8-2020==1.7.0
flake8-aaa==0.14.1
flake8-annotations==3.0.1
flake8-annotations-complexity==0.0.7
flake8-annotations-coverage==0.0.6
flake8-bandit==4.1.1
flake8-black==0.3.6
flake8-blind-except==0.2.1
flake8-breakpoint==1.1.0
flake8-broken-line==0.6.0
flake8-bugbear==23.3.12
flake8-builtins==1.5.3
flake8-class-attributes-order==0.1.3
flake8-coding==1.3.2
flake8-cognitive-complexity==0.1.0
flake8-comments==0.1.2
flake8-comprehensions==3.12.0
flake8-debugger==4.1.2
flake8-django==1.2
flake8-docstrings==1.7.0
flake8-encodings==0.5.0.post1
flake8-eradicate==1.4.0
flake8-executable==2.1.3
flake8-expression-complexity==0.0.11
flake8-fastapi==0.7.0
flake8-fixme==1.1.1
flake8-functions==0.0.8
flake8-functions-names==0.4.0
flake8-future-annotations==0.0.5
flake8-helper==0.2.1
flake8-isort==6.0.0
flake8-literal==1.3.0
flake8-logging-format==0.9.0
flake8-markdown==0.4.0
flake8-mutable==1.2.0
flake8-no-pep420==2.4.0
flake8-noqa==1.3.1
flake8-pie==0.16.0
flake8-plugin-utils==1.3.2
flake8-pyi==22.11.0
flake8-pylint==0.2.0
flake8-pytest-style==1.7.2
flake8-quotes==3.3.2
flake8-rst-docstrings==0.3.0
flake8-secure-coding-standard==1.4.0
flake8-string-format==0.3.0
flake8-tidy-imports==4.8.0
flake8-typing-imports==1.14.0
flake8-use-fstring==1.4
flake8-use-pathlib==0.3.0
flake8-useless-assert==0.4.4
flake8-variables-names==0.0.5
flake8-warnings==0.4.0
flake8_simplify==0.20.0
gitdb==4.0.10
GitPython==3.1.31
grpcio==1.51.1
grpcio-tools==1.51.1
gunicorn==20.1.0
hypothesis==6.75.2
hypothesmith==0.1.9
idna==3.4
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
jupyterlab-flake8==0.7.1
lark-parser==0.12.0
lazy-object-proxy==1.9.0
libcst==0.4.9
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
mr-proper==0.0.7
mypy-extensions==1.0.0
natsort==8.3.1
openapi-codec==1.3.2
packaging==23.1
pandas-vet==0.2.3
pathspec==0.10.3
pbr==5.11.1
pep8-naming==0.13.3
platformdirs==3.5.0
pluggy==1.0.0
protobuf==4.22.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybetter==0.4.1
pycln==2.1.3
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.7
pydocstyle==6.3.0
pyemojify==0.2.0
pyflakes==2.5.0
Pygments==2.15.1
pylint==2.17.3
pyproject_api==1.5.1
pyrsistent==0.19.3
pytest==7.3.1
pytest-cov==4.0.0
pytest-sugar==0.9.7
python-decouple==3.6
python-dev-tools==2023.3.24
pytz==2023.3
pyupgrade==3.3.2
PyYAML==6.0
removestar==1.3.1
requests==2.30.0
restructuredtext-lint==1.4.0
rich==13.3.5
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
simplejson==3.18.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==6.2.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
ssort==0.11.6
starlette==0.26.1
stdlib-list==0.8.0
stevedore==5.0.0
swagger-spec-validator==3.0.3
termcolor==2.3.0
tokenize-rt==5.0.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
tox==4.5.1
tox-travis==0.12
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
untokenize==0.1.1
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.23.0
wrapt==1.13.3
yarg==0.1.9
